# Use std::equal instead of string_view == in begins_with.