This has a slightly better codgen - https://godbolt.org/z/z3onK91bT and supports [execution policy](https://en.cppreference.com/w/cpp/algorithm/execution_policy_tag_t) in case sequential comparisons are not sufficient.