@lemire 2023-08-21
This PR appears beneficial when using GCC on x64.

Main branch (GCC 11, Icelake):

```
--------------------------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------
BasicBench_AdaURL_href              37593177 ns     37546287 ns           19 GHz=3.18835 cycle/byte=13.7422 cycles/url=1.19364k instructions/byte=38.0345 instructions/cycle=2.76771 instructions/ns=8.82444 instructions/url=3.30364k ns/url=374.375 speed=231.397M/s time/byte=4.32158ns time/url=375.369ns url/s=2.66405M/s
BasicBench_AdaURL_aggregator_href   23846082 ns     23817000 ns           29 GHz=3.18838 cycle/byte=8.72323 cycles/url=757.693 instructions/byte=26.0036 instructions/cycle=2.98096 instructions/ns=9.50443 instructions/url=2.25865k ns/url=237.642 speed=364.785M/s time/byte=2.74134ns time/url=238.11ns url/s=4.19973M/s
```


This PR (GCC 11, Icelake):

```
--------------------------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------
BasicBench_AdaURL_href              37194226 ns     37151914 ns           19 GHz=3.18876 cycle/byte=13.6456 cycles/url=1.18524k instructions/byte=37.2856 instructions/cycle=2.73243 instructions/ns=8.71308 instructions/url=3.2386k ns/url=371.694 speed=233.853M/s time/byte=4.27619ns time/url=371.426ns url/s=2.69232M/s
BasicBench_AdaURL_aggregator_href   23028129 ns     22999254 ns           30 GHz=3.1875 cycle/byte=8.46279 cycles/url=735.072 instructions/byte=25.2896 instructions/cycle=2.98833 instructions/ns=9.5253 instructions/url=2.19663k ns/url=230.611 speed=377.755M/s time/byte=2.64722ns time/url=229.935ns url/s=4.34905M/s
```