@ttsugriy 2023-08-21
thank you for running the benchmarks and accepting the PR, @lemire ! Looks like your measurements confirm a single instruction/byte reduction :)