@lemire 2023-08-21
@ttsugriy 

This PR reduces the instruction count by 2% and improves the speed by 1% in one specific benchmark. That's not huge, but these gains add up over time. Of course, I am betting that this good result generalizes, but we can also count on the assembly output which does look favorable.